remove global preferences from Imps.ProviderSettings
Imps.ProviderSettings is for account profiles, and is not the place for
global settings. These global preferences are not sensitive, so they do not
need to be stored encrypted.  Instead, the global prefs are stored in a
standard SharedPreferences file.  A static utility class is used to make
accessing the prefs from the code dead simple.

This also purges the unused "background image" and "show grid" preferences.