always use internal data store for chat logs and media
The code for handling storing the IOCipher container file on the SD card is
complicated and brittle.  There are all sorts of crashy edge cases. This
moves the IOCipher store to always be in the internal storage.  If a user
is running out of space on their /data partition, they can move the entire
Zom app to the SD Card aka External Storage.