set versionName based on git release tag
This makes it easy to tell which debug build a device is running, since the
versionName now automatically describes the exact commit that was built,
based on `git describe`, e.g.:  15.0.1-BETA-8-33-gc2e8e8a

For release builds, i.e. builds from commits that are tagged as releases,
the versionName will be just the tag name: 15.0.1-BETA-9.  So the release
tag must be created before making the release build.

This also moves versionCode from the AndroidManifest.xml to build.gradle
just to keep the two versions together.  That part is optional.